requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers>=2.7.0
faiss-cpu==1.12.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
arxiv>=1.4.8
python-multipart>=0.0.6
pydantic>=2.5.0
python-crontab>=3.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
llama-cpp-python>=0.2.0
huggingface-hub>=0.20.0
